move cli.rs into sand module
