+++ src/sand.rs
@@ -1,3 +1,5 @@
+pub mod cli;
+
 pub const VERSION: &str = "Sand v0.3.0: rewrite it in Rust";
 
 // pub enum Message {
